backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.115.0
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.1
h11==0.16.0
huggingface-hub==0.25.2
idna==3.11
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
packaging==26.0
pillow==12.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
PyYAML==6.0.3
regex==2024.9.11
requests==2.32.5
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
setuptools==70.2.0
SQLAlchemy==2.0.46
starlette==0.38.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.20.1
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.0
